pyproject.toml
[tool.poetry]
name = "satute"
version = "0.1.0"
description = ""
authors = ["CIBIV"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scipy = "^1.13.0"
numpy = "^1.26.4"
biopython = "^1.83"
ete3 = "^3.1.3"
pandas = "^2.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[tool.poetry.scripts]
satute = "satute.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
